Don't duplicate a whole library.
